requirements.txt
asgiref==3.8.1
boto3==1.26.0
botocore==1.29.0
certifi==2024.12.14
charset-normalizer==3.4.0
Django==3.2.20
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.10
jmespath==1.0.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
